query in main thread, send data to workers
